# docs/swagger.yaml
definitions:
...
  api_types.CalculationItemResponse:
    properties:
      coefficient:
        type: number
      final_dividend:
        description: omitempty скроет поле, если оно null
        type: number
      fine:
        type: number
      shareholder:
        $ref: '#/definitions/api_types.ShareholderResponse'
    type: object
  api_types.CalculationItemUpdateRequest:
    properties:
      coefficient:
        type: number
      fine:
        type: number
    type: object
...
  api_types.ShareholderRequest:
    properties:
      description:
        type: string
      name:
        type: string
      share:
        type: number
    type: object
  api_types.ShareholderResponse:
    properties:
      description:
        type: string
      id:
        type: integer
      image_url:
        description: Используем *string для nullable
        type: string
      name:
        type: string
      share:
        type: number
    type: object
...
info:
  contact: {}
paths:
...
  /dividend-calculations/{id}/shareholders/{shareholder_id}:
    delete:
      description: Removes a shareholder from the current user's draft calculation.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder ID
        in: path
        name: shareholder_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a shareholder from a draft calculation
      tags:
      - M-M
    put:
      consumes:
      - application/json
      description: Changes coefficient and fine for a shareholder within the current
        user's draft.
      parameters:
      - description: Calculation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder ID
        in: path
        name: shareholder_id
        required: true
        type: integer
      - description: Parameters to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/api_types.CalculationItemUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.CalculationItemResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not creator or not a draft)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a shareholder's parameters in a draft calculation
      tags:
      - M-M
...
  /shareholders:
    get:
      description: Get all shareholders with optional name filter. Publicly accessible.
      parameters:
      - description: Filter by shareholder name (case-insensitive)
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/api_types.ShareholderResponse'
            type: array
      summary: Get list of shareholders
      tags:
      - shareholders
    post:
      consumes:
      - application/json
      description: Add a new shareholder to the database. Requires moderator rights.
      parameters:
      - description: Shareholder object
        in: body
        name: shareholder
        required: true
        schema:
          $ref: '#/definitions/api_types.ShareholderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new shareholder (Moderator only)
      tags:
      - shareholders
  /shareholders/{id}:
    delete:
      description: Logically delete a shareholder by its ID. Requires moderator rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a shareholder (Moderator only)
      tags:
      - shareholders
    get:
      description: Get details of a single shareholder by its ID. Publicly accessible.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Shareholder not found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a shareholder by ID
      tags:
      - shareholders
    put:
      consumes:
      - application/json
      description: Update an existing shareholder's data. Requires moderator rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Shareholder object
        in: body
        name: shareholder
        required: true
        schema:
          $ref: '#/definitions/api_types.ShareholderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api_types.ShareholderResponse'
        "400":
          description: Invalid ID or request body
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a shareholder (Moderator only)
      tags:
      - shareholders
  /shareholders/{id}/add-to-draft:
    post:
      description: Adds a shareholder to the current user's draft calculation.
      parameters:
      - description: Shareholder ID to add
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID format
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Add a shareholder to the draft calculation
      tags:
      - shareholders
  /shareholders/{id}/image:
    post:
      consumes:
      - multipart/form-data
      description: Upload an image and associate it with a shareholder. Requires moderator
        rights.
      parameters:
      - description: Shareholder ID
        in: path
        name: id
        required: true
        type: integer
      - description: Image file
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image URL
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID or file
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden (not a moderator)
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Upload an image for a shareholder (Moderator only)
      tags:
      - shareholders
...
swagger: "2.0"